Add support for 64bit integers to the public API.
This CL introduces 64bit indexing into the public API. The rationale
for these changes is as follows:

1. The motivation for introducing 64bit integers in Ceres Solver
   is that we wish to solve larger problems. The limiting factor
   here are the internal data structures in Ceres Solver which
   hold and solve the Gauss-Newton system. So the public facing
   changes are made so as to enable large problems.
2. We also want to minimize the breakages caused by the API changes.
   As a result, we are not changing the API for CostFunction since
   the maximum size of an individual parameter block or the number
   of residuals of a single residual block are not limiting factors.

As a result the changes in this CL are only to objects/methods which
limit the size of the problem being constructed and solved. The changes
to the internals are minimal in this CL, they are limited to making
the changes to the public API work with the internals.

https://github.com/ceres-solver/ceres-solver/issues/960

+++ examples/rosenbrock_analytic_diff.cc
@@ -50,7 +50,7 @@ class Rosenbrock final : public ceres::FirstOrderFunction {
     return true;
   }
 
-  int NumParameters() const override { return 2; }
+  int64_t NumParameters() const override { return 2; }
 };
 
 int main(int argc, char** argv) {
